requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
asttokens==3.0.0
asyncssh==2.19.0
atpublic==5.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
DAWG-Python==0.7.2
debugpy==1.8.12
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
dpath==2.2.0
dulwich==0.22.7
dvc==3.59.0
dvc-data==3.16.9
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.2.0
fastapi==0.115.8
filelock==3.17.0
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
grandalf==0.8
gto==1.7.2
h11==0.14.0
hydra-core==1.3.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.32.0
iterative-telemetry==0.0.9
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.4.2
markdown-it-py==3.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
nvidia-nccl-cu12==2.25.1
omegaconf==2.3.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pygit2==1.17.0
Pygments==2.19.1
pygtrie==2.5.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
razdel==0.5.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
scikit-learn==1.6.1
scipy==1.15.1
scmrepo==3.3.10
semver==3.0.4
setuptools==75.8.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.1
stack-data==0.6.3
starlette==0.45.3
tabulate==0.9.0
threadpoolctl==3.5.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
wordcloud==1.9.4
xgboost==2.1.4
yarl==1.18.3
zc.lockfile==3.0.post1
